Fix fallback behavior
Versions that are _lower_ than the supported versions should not return
a higher version's module. Each version string will be compared to the
supported versions before attempting a load. If the version is greater
than or equal to the supported version the module will be loaded, else
the loop will continue. If no matches are found an exception will be
raised as usual.